% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\docType{data}
\name{bomregions}
\alias{bomregions}
\title{sample dataset for use in vignettes}
\description{
This is a dataset formerly part of the DAAG. We reproduce here for use in a vignette, since that package is no longer available on CRAN.
}
\references{
\url{https://cran.r-project.org/src/contrib/Archive/DAAG/DAAG_0.5.8.tar.gz}
}
\author{
John Maindonald and W. John Braun
}
\keyword{data}
