% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs_framework.R
\name{CreateModelFuns}
\alias{CreateModelFuns}
\title{creates a list of model functions based on parametrization}
\usage{
CreateModelFuns(paramsList)
}
\arguments{
\item{paramsList}{a names list containing the non default parameters for a model class}
}
\value{
a list of functions
}
\description{
This function serves as a wrapper for the Function Factory FunGen to generate a multiple of forecasting functions
}
