% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs_framework.R
\name{CreateWalkForwardFuns}
\alias{CreateWalkForwardFuns}
\title{split the time series and creates a list tree as the main object for further operations}
\usage{
CreateWalkForwardFuns(lTrain, lValid = 1)
}
\arguments{
\item{lTrain}{length of training intervall in a walk forward validation}

\item{lValid}{length of validation data for backtesting}
}
\value{
all splits in form of a list "out[[name]][[train]]" und "out[[name]][[valdation]]"
}
\description{
split the time series and creates a list tree as the main object for further operations
}
