% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs_framework.R
\name{DoPrediction}
\alias{DoPrediction}
\title{Do the prediction for training based on the list-tree structure and the modelFitFuns}
\usage{
DoPrediction(splitDat = splitDat, modelFitFuns = modelFitFuns, ValiLength = 1)
}
\arguments{
\item{splitDat}{list-tree}

\item{modelFitFuns}{list of generated model funs}

\item{ValiLength}{length of Validation data for backtesting}
}
\value{
the list tree with addional information
}
\description{
This function enriches the list tree structure with the forecast results
Step 1: The models in modelFitFuns are fitted to alle the training splits
Step 2: The forecast is performed
Step 3: Results are attached to the list tree
}
