% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs_framework.R
\name{forecast.lm}
\alias{forecast.lm}
\title{this is  a wrapper around the lm function in order to generate a result similar to forecast(ArimaObject)}
\usage{
\method{forecast}{lm}(fittedModel, ValiLength = 1)
}
\arguments{
\item{ValiLength}{length of Validation data for backtesting}

\item{Model}{a fitted model from a lm call}
}
\value{
a list containing the forecast result
}
\description{
Note: in the function are two hidden parameters for calculating the s70 and s85 confidence intervals
}
\keyword{internal}
