% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs_framework.R
\name{forecastGen}
\alias{forecastGen}
\title{Function that serves as a warpper to forecasts}
\usage{
forecastGen(fittedModel, ValiLength = 1)
}
\arguments{
\item{Model}{a fitted Model}

\item{ValiLenght}{length of Validation data for backtesting}
}
\value{
forecast object
}
\description{
This is to create a unique way to create a forecast object. The forecast object are similar independently of the class of the fitted model
}
