% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs_framework.R
\name{FunGen}
\alias{FunGen}
\title{Function generator that returns a model functions for Model fitting}
\usage{
FunGen(params)
}
\arguments{
\item{params}{a named listwith one element containing the non-default params}
}
\value{
a function
}
\description{
This is to create a unique way to define functions that can be called exactly in the same way
The returned function is such, that the output can be used by the subsequent call to forecast
}
