% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model

%Two new commands to help with formatting
\newcommand{\methodtable}{\bold{\cr #1 \cr} \tabular{lcl}}
\newcommand{\methoditem}{\emph{#1} \tab-\tab #2\cr}
\newcommand{\methodname}{\subsection{#1}}
\docType{class}
\name{Model}
\alias{Model}
\title{Model}
\description{An abstract class for making predictions.  This class contains only what's necessary for data based modeling.  For the most part, the other model classes in this package are there to make extending this class easier.  You should extend this class directly if you are making a model which does not forecast, and none of the convenience classes seem appropriate.}
\seealso{
}
\section{Methods}{
\methodname{fit(data)}{
This method \bold{must} be extended.  Get the model ready to predict.
\describe{\item{\emph{Arguments}}{\tabular{lcl}{
\methoditem{data}{The data to fit the model to.}
}}}}
\methodname{predict(newdata)}{
This method \bold{must} be extended.  Predict using the model
\describe{\item{\emph{Arguments}}{\tabular{lcl}{
\methoditem{newdata}{the data to predict.}
}}}}
}
