% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs_framework.R
\name{needsTSconversion}
\alias{needsTSconversion}
\title{Helper to detect the neccessity for a conversion}
\usage{
needsTSconversion(inFun)
}
\arguments{
\item{inFun}{a function}
}
\value{
flag
}
\description{
This is a helper function to determine based on the prediction function if a conversion to time-series is necessary
For instance forecast::arima with seasonality expects a timesseries input with given frequency
}
