% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs_framework.R
\name{plotSingleModelSingleAccount}
\alias{plotSingleModelSingleAccount}
\title{Plots the time series for a given account and model and split}
\usage{
plotSingleModelSingleAccount(
  accountName,
  splitDat,
  modelName = "lastval",
  split = 1
)
}
\arguments{
\item{accountName}{name of account to be plotted}

\item{splitDat}{fitted list tree}

\item{modelName}{name of the model result to be plotted as defined in the modelParams}

\item{split}{index of the split to be plotted}
}
\description{
This function identifies the correct account in the tree-list and displays all forecasts with confidence intervalls in a single plot
}
